requirements.txt
# Thin UI (install after ComfyUI deps; see setup.ps1)
gradio>=4.44.0
websocket-client>=1.8.0
requests>=2.31.0
huggingface_hub>=0.27.0
transformers>=4.44.0
soundfile>=0.12.1
psutil>=6.0.0
